refactor(linter): split locales rules domain (Task 2.2)
* split 549 lines of  into logical submodules
* extract , , ,  modules
* establish  orchestration module
